% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rotn_datasets}
\alias{rotn_datasets}
\title{List OTN datasets}
\usage{
rotn_datasets()
}
\value{
A \code{data.frame} with three columns: each dataset's
    \code{id}, \code{url}, and \code{title}. These are taken from
    the OTN's API (\url{https://opentraits.org/datasets.json}).
}
\description{
Download all datasets currently indexed in The Open Trait Network
}
\examples{
datasets <- rotn_datasets()
head(datasets)
}
